feat: Implement Kubernetes Backend (#68)
* Implement Kubernetes Backends


* add k8s_test to test kubernetes backend


* run ruff format


* address reviews


* Enabling KubernetesBackendConfig to use arbitrary types

Kubernetes client configuration isn't supported as pydantic core
schema so we need to ask pydantic to allow arbitrary types.


* Remove BaseBackend and use BaseModel


* Rename base backend to ExecutionBackend


* fix api trainer_client order


* move delete_job func to the end of trainer_client


---------
