+++ kubeflow/trainer/backends/kubernetes/backend_test.py
@@ -18,6 +18,7 @@ Unit tests for the TrainerClient class in the Kubeflow Trainer SDK.
 This module uses pytest and unittest.mock to simulate Kubernetes API interactions.
 It tests TrainerClient's behavior across job listing, resource creation etc
 """
+
 import datetime
 import multiprocessing
 import random
@@ -28,11 +29,13 @@ from typing import Any, Dict, Optional, Type
 from unittest.mock import Mock, patch
 
 import pytest
-from kubeflow.trainer import TrainerClient
+from kubeflow_trainer_api import models
+
 from kubeflow.trainer.constants import constants
 from kubeflow.trainer.types import types
 from kubeflow.trainer.utils import utils
-from kubeflow_trainer_api import models
+from kubeflow.trainer.backends.kubernetes.types import KubernetesBackendConfig
+from kubeflow.trainer.backends.kubernetes.backend import KubernetesBackend
 
 
 @dataclass
@@ -76,31 +79,33 @@ TRAIN_JOB_WITH_CUSTOM_TRAINER = "train-job-with-custom-trainer"
 @pytest.fixture
 def trainer_client(request):
     """Provide a TrainerClient with mocked Kubernetes APIs."""
-    with patch("kubernetes.config.load_kube_config", return_value=None), patch(
-        "kubernetes.client.CustomObjectsApi",
-        return_value=Mock(
-            create_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
-            patch_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
-            delete_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
-            get_namespaced_custom_object=Mock(
-                side_effect=get_namespaced_custom_object_response
+    with (
+        patch("kubernetes.config.load_kube_config", return_value=None),
+        patch(
+            "kubernetes.client.CustomObjectsApi",
+            return_value=Mock(
+                create_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
+                patch_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
+                delete_namespaced_custom_object=Mock(side_effect=conditional_error_handler),
+                get_namespaced_custom_object=Mock(
+                    side_effect=get_namespaced_custom_object_response
+                ),
+                get_cluster_custom_object=Mock(side_effect=get_cluster_custom_object_response),
+                list_namespaced_custom_object=Mock(
+                    side_effect=list_namespaced_custom_object_response
+                ),
+                list_cluster_custom_object=Mock(side_effect=list_cluster_custom_object),
             ),
-            get_cluster_custom_object=Mock(
-                side_effect=get_cluster_custom_object_response
-            ),
-            list_namespaced_custom_object=Mock(
-                side_effect=list_namespaced_custom_object_response
-            ),
-            list_cluster_custom_object=Mock(side_effect=list_cluster_custom_object),
         ),
-    ), patch(
-        "kubernetes.client.CoreV1Api",
-        return_value=Mock(
-            list_namespaced_pod=Mock(side_effect=list_namespaced_pod_response),
-            read_namespaced_pod_log=Mock(side_effect=mock_read_namespaced_pod_log),
+        patch(
+            "kubernetes.client.CoreV1Api",
+            return_value=Mock(
+                list_namespaced_pod=Mock(side_effect=list_namespaced_pod_response),
+                read_namespaced_pod_log=Mock(side_effect=mock_read_namespaced_pod_log),
+            ),
         ),
     ):
-        yield TrainerClient()
+        yield KubernetesBackend(KubernetesBackendConfig())
 
 
 # --------------------------
@@ -231,7 +236,7 @@ def get_custom_trainer(
             "torch numpy \n\nread -r -d '' SCRIPT << EOM\n\nfunc=lambda: "
             'print("Hello World"),\n\n<lambda>('
             "{'learning_rate': 0.001, 'batch_size': 32})\n\nEOM\nprintf \"%s\" "
-            '"$SCRIPT" > "trainer_client_test.py"\ntorchrun "trainer_client_test.py"',
+            '"$SCRIPT" > "backend_test.py"\ntorchrun "backend_test.py"',
         ],
         numNodes=2,
         env=env,
@@ -294,9 +299,7 @@ def get_namespaced_custom_object_response(*args, **kwargs):
     if args[2] == RUNTIME or args[4] == RUNTIME:
         raise RuntimeError()
     if args[3] == TRAIN_JOBS:  # TODO: review this.
-        mock_thread.get.return_value = add_status(
-            create_train_job(train_job_name=args[4])
-        )
+        mock_thread.get.return_value = add_status(create_train_job(train_job_name=args[4]))
 
     return mock_thread
 
@@ -466,9 +469,7 @@ def create_cluster_training_runtime(
                     name=name,
                     namespace=namespace,
                 ),
-                spec=models.JobsetV1alpha2JobSetSpec(
-                    replicatedJobs=[get_replicated_job()]
-                ),
+                spec=models.JobsetV1alpha2JobSetSpec(replicatedJobs=[get_replicated_job()]),
             ),
         ),
     )
@@ -636,9 +637,7 @@ def test_list_runtimes(trainer_client, test_case):
         assert test_case.expected_status == SUCCESS
         assert isinstance(runtimes, list)
         assert all(isinstance(r, types.Runtime) for r in runtimes)
-        assert [asdict(r) for r in runtimes] == [
-            asdict(r) for r in test_case.expected_output
-        ]
+        assert [asdict(r) for r in runtimes] == [asdict(r) for r in test_case.expected_output]
 
     except Exception as e:
         assert type(e) is test_case.expected_error
@@ -755,12 +754,8 @@ def test_get_runtime_packages(trainer_client, test_case):
                 train_job_name=TRAIN_JOB_WITH_CUSTOM_TRAINER,
                 train_job_trainer=get_custom_trainer(
                     env=[
-                        models.IoK8sApiCoreV1EnvVar(
-                            name="TEST_ENV", value="test_value"
-                        ),
-                        models.IoK8sApiCoreV1EnvVar(
-                            name="ANOTHER_ENV", value="another_value"
-                        ),
+                        models.IoK8sApiCoreV1EnvVar(name="TEST_ENV", value="test_value"),
+                        models.IoK8sApiCoreV1EnvVar(name="ANOTHER_ENV", value="another_value"),
                     ],
                 ),
             ),
@@ -800,9 +795,7 @@ def test_train(trainer_client, test_case):
     print("Executing test:", test_case.name)
     try:
         trainer_client.namespace = test_case.config.get("namespace", DEFAULT_NAMESPACE)
-        runtime = trainer_client.get_runtime(
-            test_case.config.get("runtime", TORCH_RUNTIME)
-        )
+        runtime = trainer_client.get_runtime(test_case.config.get("runtime", TORCH_RUNTIME))
 
         train_job_name = trainer_client.train(
             runtime=runtime, trainer=test_case.config.get("trainer", None)
@@ -910,9 +903,7 @@ def test_list_jobs(trainer_client, test_case):
         assert test_case.expected_status == SUCCESS
         assert isinstance(jobs, list)
         assert len(jobs) == 2
-        assert [asdict(j) for j in jobs] == [
-            asdict(r) for r in test_case.expected_output
-        ]
+        assert [asdict(j) for j in jobs] == [asdict(r) for r in test_case.expected_output]
 
     except Exception as e:
         assert type(e) is test_case.expected_error
@@ -1037,9 +1028,7 @@ def test_wait_for_job_status(trainer_client, test_case):
         assert test_case.expected_status == SUCCESS
         assert isinstance(job, types.TrainJob)
         # Job status should be in the expected set.
-        assert job.status in test_case.config.get(
-            "status", {constants.TRAINJOB_COMPLETE}
-        )
+        assert job.status in test_case.config.get("status", {constants.TRAINJOB_COMPLETE})
 
     except Exception as e:
         assert type(e) is test_case.expected_error
